chore(deps): bump Vortice.XAudio2 from 3.5.0 to 3.6.2

+++ src/Beutl.Engine/Audio/Platforms/XAudio2/XAudioBuffer.cs
@@ -45,7 +45,7 @@ public sealed unsafe class XAudioBuffer : IDisposable
         Format = format;
         SizeInBytes = sizeInBytes;
         Buffer.AudioDataPointer = (nint)_stream;
-        Buffer.AudioBytes = SizeInBytes;
+        Buffer.AudioBytes = (uint)SizeInBytes;
     }
 
     public void Dispose()
